% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{as.data.frame,Data-method}
\alias{as.data.frame,Data-method}
\title{Casting \code{Data} object to data.frame.}
\usage{
\S4method{as.data.frame}{Data}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{any \R object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link{data.frame}(*, check.names = !optional)}.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Casting \code{Data} object to data.frame.
}
