% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble,Data-method}
\title{Generic as_tibble}
\usage{
as_tibble(x)

\S4method{as_tibble}{Data}(x)
}
\arguments{
\item{x}{\code{Data} object.}
}
\description{
Generic as_tibble
}
