% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{convert_fields_to_date}
\alias{convert_fields_to_date}
\title{Internal method to convert fields to dates}
\usage{
convert_fields_to_date(obj, word)
}
\arguments{
\item{obj}{\code{RawData} object}

\item{word}{character Key word suggesting date format}
}
\description{
Internal method to convert fields to dates
}
