% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-functions.R
\name{create_api_call}
\alias{create_api_call}
\title{This is an internal function to create the API query}
\usage{
create_api_call(query, api_key, category, db, count_var = "", limit = 0,
  skip = 0)
}
\arguments{
\item{query}{character Query string.}

\item{api_key}{character Private api key.}

\item{category}{character Category like 'device'.}

\item{db}{character Database like 'event'.}

\item{count_var}{character Name of counting field, if raw keep "".}

\item{limit}{integer   For raw, set limit of records to be fetch <5100.}

\item{skip}{integer   Skip * 100 records.}
}
\description{
This is an internal function to create the API query
}
