% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{dim,Data-method}
\alias{dim,Data-method}
\title{\code{dim} method for the \code{Data} class.}
\usage{
\S4method{dim}{Data}(x)
}
\arguments{
\item{x}{an \R object, for example a matrix, array or data frame.}
}
\description{
\code{dim} method for the \code{Data} class.
}
