% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{fetch}
\alias{fetch}
\alias{fetch,CountQuery-method}
\alias{fetch,RawQuery-method}
\title{Generic function for fetching data from the API based on the \code{Query}
object.}
\usage{
fetch(.Object, ...)

\S4method{fetch}{CountQuery}(.Object, json = FALSE, quiet = FALSE)

\S4method{fetch}{RawQuery}(.Object, json = FALSE, quiet = FALSE)
}
\arguments{
\item{.Object}{An instance of the \code{Query} subclass.}

\item{...}{additional parameters like json, see openfda.}

\item{json}{logical if TRUE, store the raw JSON response in the object.}

\item{quiet}{logical if TRUE, no warning messages are generated.}
}
\value{
An object representing the fetching results and including the original
\code{Query} object. The class returned is specific to the combination of
the category and database interrogated.
}
\description{
Generic function for fetching data from the API based on the \code{Query}
object.
}
