% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{fetch_json}
\alias{fetch_json}
\title{Function to fetch JSON form API.}
\usage{
fetch_json(query_api)
}
\arguments{
\item{query_api}{character Representing the query to API.}
}
\description{
Function to fetch JSON form API
}
