% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-functions.R
\name{firstup}
\alias{firstup}
\title{This is an internal function to capitalize the first letter.}
\usage{
firstup(x)
}
\arguments{
\item{x}{Input string.}
}
\description{
This is an internal function to capitalize the first letter.
}
