% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{head,Data-method}
\alias{head,Data-method}
\title{\code{head} of the \code{Data} class.}
\usage{
\S4method{head}{Data}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
\code{head} of the \code{Data} class.
}
