% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{names,Data-method}
\alias{names,Data-method}
\title{names method for class \code{Data}.}
\usage{
\S4method{names}{Data}(x)
}
\arguments{
\item{x}{\code{Data} object.}
}
\description{
names method for class \code{Data}.
}
