% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{ncol,Data-method}
\alias{ncol,Data-method}
\title{\code{ncol} of the \code{Data} class.}
\usage{
\S4method{ncol}{Data}(x)
}
\arguments{
\item{x}{\code{Data} object.}
}
\description{
\code{ncol} of the \code{Data} class.
}
