% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{nrow,Data-method}
\alias{nrow,Data-method}
\title{\code{nrow} of the \code{Data} class.}
\usage{
\S4method{nrow}{Data}(x)
}
\arguments{
\item{x}{\code{Data} object.}
}
\description{
\code{nrow} of the \code{Data} class.
}
