% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{plot,CountData,ANY-method}
\alias{plot,CountData,ANY-method}
\title{\code{plot} of the \code{CountData} class.}
\usage{
\S4method{plot}{CountData,ANY}(x)
}
\arguments{
\item{x}{\code{CountData} object.}
}
\description{
\code{plot} of the \code{CountData} class.
}
