% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{plot,RawData,missing-method}
\alias{plot,RawData,missing-method}
\title{\code{plot} of the \code{RawData} class.}
\usage{
\S4method{plot}{RawData,missing}(x, y, i = 1, j = 2, ...)
}
\arguments{
\item{x}{\code{RawData} object.}

\item{y}{not used}

\item{i}{numeric of character specifing index for what to plot on the x-axis}

\item{j}{numeric of character specifing index for what to plot on the x-axis}

\item{...}{not used}
}
\description{
\code{plot} of the \code{RawData} class.
}
