% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{Query-class}
\alias{Query-class}
\alias{CountQuery-class}
\alias{RawQuery-class}
\title{Query class.}
\description{
\code{Query} is a virtual S4 class that represents a general query to the
openFDA API.

\code{CountQuery} is an S4 class that represent a query for count data to the
open FDA API. It inherits from \code{Query} class.

\code{RawQuery} is an S4 class that represent a query for raw data to the
openFDA API. It inherits from \code{Query} class.
}
\section{Slots}{

\describe{
\item{\code{query_date}}{POSIXct   The query creation date.}

\item{\code{query}}{character The query string provided by the user.}

\item{\code{query_api}}{noquote   The query string constructed for the API call.}

\item{\code{db}}{character openFDA database name.}

\item{\code{category}}{character openFDA category name.}

\item{\code{api_key}}{character The private key.}

\item{\code{count_var}}{character Specific slot for \code{CountQuery} class
representing the name of the counting field.}

\item{\code{limit}}{numeric Specific slot for \code{RawQuery} class representing
a maximum number of rows requested from API. openFDA API is not allowing
returning more than 5100 records.}
}}

