% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\docType{package}
\name{ropenfda}
\alias{ropenfda}
\alias{ropenfda-package}
\title{Accessing openFDA API device information.}
\description{
The \code{ropenfda} package provides classes and methods for
conveniently accessing, manipulating and visualizing data openFDA API.
}
