% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{show,CountQuery-method}
\alias{show,CountQuery-method}
\title{Show method for class \code{CountQuery}.}
\usage{
\S4method{show}{CountQuery}(object)
}
\arguments{
\item{object}{\code{CountQuery} object.}
}
\description{
Show method for class \code{CountQuery}.
}
