% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{show,Data-method}
\alias{show,Data-method}
\title{Show method for class \code{Data}.}
\usage{
\S4method{show}{Data}(object)
}
\arguments{
\item{object}{\code{Data} object.}
}
\description{
Show method for class \code{Data}.
}
