% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{show,Query-method}
\alias{show,Query-method}
\title{Show method for class \code{Query}.}
\usage{
\S4method{show}{Query}(object)
}
\arguments{
\item{object}{\code{Query} object.}
}
\description{
Show method for class \code{Query}.
}
