% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-functions.R
\name{str2api}
\alias{str2api}
\title{This is an internal function for parsing the query to the proper API grammar.}
\usage{
str2api(s)
}
\arguments{
\item{s}{character Input string.}
}
\description{
This is an internal function for parsing the query to the proper API grammar.
}
