% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\docType{methods}
\name{unfold}
\alias{unfold}
\alias{unfold,RawDeviceEvent-method}
\alias{unfold,RawDeviceClassification-method}
\alias{unfold,RawDevice510k-method}
\alias{unfold,RawDevicePma-method}
\alias{unfold,RawDeviceRegistrationlisting-method}
\alias{unfold,RawDeviceRecall-method}
\alias{unfold,RawDeviceEnforcement-method}
\alias{unfold,RawDeviceUdi-method}
\title{Method to unfold the nested raw data inside the results from the API.}
\usage{
unfold(.Object)

\S4method{unfold}{RawDeviceEvent}(.Object)

\S4method{unfold}{RawDeviceClassification}(.Object)

\S4method{unfold}{RawDevice510k}(.Object)

\S4method{unfold}{RawDevicePma}(.Object)

\S4method{unfold}{RawDeviceRegistrationlisting}(.Object)

\S4method{unfold}{RawDeviceRecall}(.Object)

\S4method{unfold}{RawDeviceEnforcement}(.Object)

\S4method{unfold}{RawDeviceUdi}(.Object)
}
\arguments{
\item{.Object}{An instance of the class representing the outcome from the API
from a specific category and database combination.}
}
\description{
These methods work only on the raw data. These data structures might be
nested data.frames. Count data is never nested. \code{unfold} methods are
customized for individual openFDA databases. These functions make assumptions
what might be important to the user how to unfold these nested data
structures. For instance the device event database results contain records
that may contain multiple patients and multiple devices per one row.
\code{unfold} method would create a cartesian product of patients and devices
with the rest of the columns copied.
}
