% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{Data-class}
\alias{Data-class}
\alias{CountData-class}
\alias{RawData-class}
\alias{RawDeviceUdi-class}
\alias{RawDeviceEvent-class}
\alias{RawDeviceClassification-class}
\alias{RawDevice510k-class}
\alias{RawDevicePma-class}
\alias{RawDeviceRegistrationlisting-class}
\alias{RawDeviceRecall-class}
\alias{RawDeviceEnforcement-class}
\alias{CountDeviceEvent-class}
\alias{CountDeviceClassification-class}
\alias{CountDevice510k-class}
\alias{CountDevicePma-class}
\alias{CountDeviceRegistrationlisting-class}
\alias{CountDeviceRecall-class}
\alias{CountDeviceEnforcement-class}
\alias{CountDeviceUdi-class}
\title{Data class.}
\description{
\code{Data} is an S4 virtual class that represents a data fetched from the
openFDA API.

\code{CountData} is an S4 class that represents a data fetched from the
open FDA API. It inherits from \code{Data} class.

\code{RawData} is an S4 class that represent a data fetched from the
open FDA API. It inherits from \code{Data} class.

\code{RawDeviceUdi} A specialized subclass representing fetched data from API.

\code{RawDeviceEvent} A specialized subclass representing fetched data from
API.

\code{RawDeviceClassification} A specialized subclass representing fetched
data from API.

\code{RawDevice510k} A specialized subclass representing fetched data from
API.

\code{RawDevicePma} A specialized subclass representing fetched data from
API.

\code{RawDeviceRegistrationlisting} A specialized subclass representing
fetched data from API.

\code{RawDeviceRecall} A specialized subclass representing fetched data from
API.

\code{RawDeviceEnforcement} A specialized subclass representing fetched data
from API.

\code{CountDeviceEvent} A specialized subclass representing fetched data from
API.

\code{CountDeviceClassification} A specialized subclass representing fetched
data from API.

\code{CountDevice510k} A specialized subclass representing fetched data from
API.

\code{CountDevicePma} A specialized subclass representing fetched data from
API.

\code{CountDeviceRegistrationlisting} A specialized subclass representing
fetched data from API.

\code{CountDeviceRecall} A specialized subclass representing fetched data
from API.

\code{CountDeviceEnforcement} A specialized subclass representing fetched
data from API.

\code{CountDeviceUdi} A specialized subclass representing fetched data
from API.
}
\section{Slots}{

\describe{
\item{\code{query_time}}{POSIXct    The time the data was fetched.}

\item{\code{result}}{data.frame The resulting data.}

\item{\code{meta}}{data.frame The meta-information such as number of total
records found in the database, error messages, license, date of the last
update of the database, etc.}

\item{\code{json}}{character  Raw JSON response from the API. Contains values
only if fetched with the option \code{json = TRUE}.}

\item{\code{query}}{CountQuery Object representing the original query.}

\item{\code{query}}{RawQuery Object representing the original query.}
}}

